rename pkg name to utils
