+++ pkg/utils/arrays.go
@@ -1,4 +1,4 @@
-package safemath
+package utils
 
 func Reverse[T any](a []T) {
 	for i, j := 0, len(a)-1; i < j; i, j = i+1, j-1 {
