+++ pkg/utils/safemath.go
@@ -1,4 +1,4 @@
-package safemath
+package utils
 
 import (
 	"math/bits"
