+++ pkg/utils/safemath_test.go
@@ -1,4 +1,4 @@
-package safemath
+package utils
 
 import (
 	"testing"
